Use old version of GHC to make use of binary cache
I tried to fix the problem of always building a lot of derivations after
introducing haskell-language-server, because my machine's performance
limit made it unable to successfully build nixvim before. The size of
nixvim derivation with haskell-language-server reached 11 GB. After this
modification, nixvim was successfully built on this machine. But I found
that in the old version of nixpkg (24.11) I was using, the versions of
derivations that were previously built manually without hitting the
binary cache were the same as the previous unstable (2c8d3f48), such as
lsp-types 2.3.0.0; all the big derivations were also the same, such as
ghc 9.6.6, haskell-language-server 2.9.0.0. When I used nix-tree to
check, I also found that the newly built nixvim did use ghc 9.6.6 and
other packages that are consistent with the current unstable. In this
case, it is unknown whether this commit has any effect. The only
explanation I can give for the inexplicable successful compilation is
that two hours ago, there was a small problem with my binary cache that
caused me to fail to pull some specific packages, and now (regardless of
whether I made the changes in this commit) they have been fixed.